Split HolidayManagerTest into performance and use case tests
