+++ jollyday-core/src/main/java/de/focus_shift/jollyday/core/HolidayManager.java
@@ -23,9 +23,6 @@ import static java.util.stream.Collectors.toUnmodifiableSet;
  * Abstract base class for all holiday manager implementations. Upon call of
  * getInstance method the implementing class will be read from the
  * jollyday.properties file and instantiated.
- *
- * @author Sven Diedrichsen
- * @version $Id: $
  */
 public abstract class HolidayManager {
